requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==23.2.1
Pillow==10.3.0
httpx==0.27.0
